Add code for People and start PeopleFeatures
`People` is a class to hold people.

`read_in_people()` reads in the people, but does not remove anyone or
update counts in the features.

`PeopleFeatures` has a reference to both `People` and
`FeatureCollection` and has code to update counts in the features and
delete people who we can be confident won't be picked.