Flesh out and test PeopleFeatures.select_person()
When someone is selected we update the "selected" and "remaining" parts
of the FeatureCollection and remove them from the PeopleFeatures copy of
People.

If we are doing address checks, then we also look for anyone with the same
address as the selected person and remove them from the copy of People,
and update "remaining" (but not "selected").