Make it possible to set the random seed
By making it a choice whether we use secrets or random module.

This is something of a hack. Mostly we want to use the `secrets` module.
But for repeatable testing we might want to set the random.seed sometimes.

So we have a global `_random_provider` which can be switched between an
instance of this class that uses the `secrets` module and an instance that
uses `random` with a seed. The switch is done by the `set_random_provider()`
function.

Then every time we want some randomness, we call `random_provider()` to get
the current version of the global.  That global has classmethods for the
random things we want to do in this library.