Split FeatureCollection into two parallel sets of classes
One just holds the min/max (and flex versions of min/max). That is the
main FeatureCollection that is used by all the algorithms.

The new SelectCollection holds the selected/remaining counts - and has a
reference to the min/max values held by the FeatureCollection.  This is
only used by the legacy selection algorithm.