Move to DataSource pattern - CSV for this commit
We now have a `SelectionData` class that orchestrates reading and
writing data.  It needs an instance that inherits from
`AbstractDataSource` to use to read and write the data for running
selection.

So rather than having a single `CSVAdapter` class that has methods for
reading from and writing to both strings and files, we now have 2 CSV
data source classes: `CSVFileDataSource` and `CSVStringDataSource`.