Add errors to the RunReport
And allow some errors to have multiple lines, that will appear on
multiple lines in the RunReport.

As part of this, clean up GSheetDataSource to not try and stuff a report
into an error, and so stop the SelectionError from storing a report.  It
only ever stored the name of the spreadsheet when we successfully opened
it, so instead we just report that name in the error message when it is
useful.