Check for duplicate rows in people data
Duplicate rows means two rows that have the same value in the column
used as the "unique id".

If any are found, they are reported. If the rows turn out to be
identical, then the duplicates are reported, the duplicates will be
ignored for selection, and selection can continue.

If any duplicate rows do not have matching data then the details will be
reported and processing will stop - the user will have to sort out the
data before continuing to do selection.