Use "feature" or "category" rather than "feature/category" in error messages
This means finding the column header used and passing that around, so it
is available for error messages around the code.

It makes the list of arguments a little long sometimes, and means we're
returning a tuple rather than a single thing here and there. Might
refactor later, but will leave it for now.

And then quite a lot of tests need updating