Only retry for legacy algorithm failures
No point retrying for the others, that use the solver. They will always
find a solution if there is one.

To do this, we make it possible to determine if a raised error is
intended to be retry-able and only retry for those errors. Then we only
mark errors as retry-able in the legacy algorithm code.